refactor(evm): extract shared OpStackEvmClient for Optimism + Base
Optimism and Base are both OP Stack L2s and had identical implementations
of getCurrentGasCostForCoinTransaction, getCurrentGasCostForTokenTransaction,
getTxActualFee and the l2Provider getter (all using @eth-optimism/sdk).

Move them into a shared abstract OpStackEvmClient. OptimismClient and
BaseClient now only contain their bridge-specific logic (CrossChainMessenger).

Each had its own local OptimismTransactionReceipt / BaseTransactionReceipt
interface with the same fields; collapsed into one OpStackTransactionReceipt.
